Workflow: single-file update
The implementation is different w.r.t. the previous commit, but the focus is the same, and the obtained workflow is more similar to the original version.
Note: setting "python-version: 3.8" in miniconda setup  for ubuntu-latest result in "mujoco.FatalError: gladLoadGL error", even if "MUJOCO_GL=egl" is set as env variable.